pyproject.toml
[tool.poetry]
name = "agoa"
version = "0.1.0"
description = "Airport Ground Operations Application"
authors = ["Ali Daoudi"]
readme = "README.md"
packages = [{ include = "agoa" }]

[tool.poetry.dependencies]
python = ">=3.12"
django = ">=5.1.6,<6.0.0"
djangorestframework = ">=3.15.2,<4.0.0"
djangorestframework-simplejwt = ">=5.4.0,<6.0.0"
drf-spectacular = ">=0.27.1,<1.0.0"
django-cors-headers = ">=4.7.0,<5.0.0"
coverage = "^7.6.12"

[tool.poetry.group.dev.dependencies]
black = ">=25.1.0,<26.0.0"
pytest = ">=8.3.4,<9.0.0"
pytest-django = ">=4.10.0,<5.0.0"
mypy = ">=1.15.0,<2.0.0"
django-stubs = ">=5.1.3,<6.0.0"

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest]
DJANGO_SETTINGS_MODULE = "agoa.settings"
python_files = ["test_*.py", "*_test.py", "tests.py"]
addopts = "-v"
